Update routes and update app.ts
