+++ src/routes/RoutesIndex.ts
@@ -1,0 +1,8 @@
+import Index from "@/controllers/Index";
+import express from "express";
+const router = express.Router();
+
+/* GET home page. */
+router.get("/", Index.get);
+
+export default router;
