+++ src/routes/RoutesUser.ts
@@ -1,0 +1,12 @@
+import User from "@/controllers/User";
+import express from "express";
+const router = express.Router();
+
+/* user page. */
+router.get("/user", User.get);
+router.get("/user/:id", User.get);
+router.post("/user", User.post);
+router.patch("/user/:id", User.patch);
+router.delete("/user/:id", User.delete);
+
+export default router;
